fix: a problem with duplicate featured publication
